# Fixed #28518 -- Improved performance of loading geometries from DB.