--- a/django/contrib/gis/db/backends/oracle/operations.py
+++ b/django/contrib/gis/db/backends/oracle/operations.py
@@ -16,6 +16,8 @@ from django.contrib.gis.db.backends.oracle.adapter import OracleSpatialAdapter
 from django.contrib.gis.db.backends.utils import SpatialOperator
 from django.contrib.gis.db.models import aggregates
 from django.contrib.gis.geometry.backend import Geometry
+from django.contrib.gis.geos import GEOSGeometry
+from django.contrib.gis.geos.prototypes.io import wkb_r
 from django.contrib.gis.measure import Distance
 from django.db.backends.oracle.operations import DatabaseOperations
 
@@ -85,7 +87,7 @@ class OracleOperations(BaseSpatialOperations, DatabaseOperations):
     # However, this adversely affects performance (i.e., Java is called
     # to convert to WKT on every query).  If someone wishes to write a
     # SDO_GEOMETRY(...) parser in Python, let me know =)
-    select = 'SDO_UTIL.TO_WKTGEOMETRY(%s)'
+    select = 'SDO_UTIL.TO_WKBGEOMETRY(%s)'
 
     gis_operators = {
         'contains': SDOOperator(func='SDO_CONTAINS'),
@@ -112,24 +114,12 @@ class OracleOperations(BaseSpatialOperations, DatabaseOperations):
     def geo_quote_name(self, name):
         return super().geo_quote_name(name).upper()
 
-    def get_db_converters(self, expression):
-        converters = super().get_db_converters(expression)
-        internal_type = expression.output_field.get_internal_type()
-        geometry_fields = (
-            'PointField', 'GeometryField', 'LineStringField',
-            'PolygonField', 'MultiPointField', 'MultiLineStringField',
-            'MultiPolygonField', 'GeometryCollectionField',
-        )
-        if internal_type in geometry_fields:
-            converters.append(self.convert_textfield_value)
-        return converters
-
     def convert_extent(self, clob):
         if clob:
             # Generally, Oracle returns a polygon for the extent -- however,
             # it can return a single point if there's only one Point in the
             # table.
-            ext_geom = Geometry(clob.read())
+            ext_geom = Geometry(memoryview(clob.read()))
             gtype = str(ext_geom.geom_type)
             if gtype == 'Polygon':
                 # Construct the 4-tuple from the coordinates in the polygon.
@@ -207,3 +197,13 @@ class OracleOperations(BaseSpatialOperations, DatabaseOperations):
         if placeholder == 'NULL':
             return []
         return super().modify_insert_params(placeholder, params)
+
+    def get_geometry_converter(self, expression):
+        read = wkb_r().read
+        srid = expression.output_field.srid
+        if srid == -1:
+            srid = None
+
+        def converter(value, expression, connection):
+            return None if value is None else GEOSGeometry(read(memoryview(value.read())), srid)
+        return converter