https://code.djangoproject.com/ticket/28518

Before:
```
In [2]: %timeit for x in City.objects.values_list('point')[:1000]: pass
64.4 ms ± 644 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
After:
```
In [3]: %timeit for x in City.objects.values_list('point')[:1000]: pass
49.8 ms ± 957 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Benchmarks are provided for PostgreSQL, will provide them for other backends if needed.