@sir-sigurd 2017-08-23
buildbot, test on xenial.